% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BoundedWeightedVoronoi}
\alias{BoundedWeightedVoronoi}
\title{Bounded Weighted Voronoi Diagram}
\usage{
BoundedWeightedVoronoi(coordX, coordY, Weights, minX, minY, maxX, maxY)
}
\arguments{
\item{coordX}{numeric vector of x coordinates}

\item{coordY}{numeric vector of y coordinates}

\item{Weights}{numeric vector of weights}

\item{minX}{coordinates of bounding rectangle}

\item{minY}{coordinates of bounding rectangle}

\item{maxX}{coordinates of bounding rectangle}

\item{maxY}{coordinates of bounding rectangle}
}
\description{
Produce a weighted and bounded Voronoi diagram. This is a wrapper for the CGAL library, see \url{http://doc.cgal.org/latest/Apollonius_graph_2/index.html} for algorithm details.
}
