% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCells.R
\name{readCells}
\alias{readCells}
\title{Format Output of CGAL Code}
\usage{
readCells(voronoiOut)
}
\arguments{
\item{voronoiOut}{a list output}
}
\description{
The readCells(). function reads the file diagram.txt.
This is the output file that the external program voronoiDiagram generates.
The file contains multiple lines for each vertex (site), so each region (cell) is represented by a block of lines in the file.
The blocks are read in the order that the sites were specified in R (by matching the site locations in R to the vertex information in the file), rather than the order that they appear in the file (which may be different).
Notice that this function handles two special cases: it is possible that a site location does not appear (as a vertex) in the diagram.txt file at all; it is also possible that a site location appears (as a vertex) in the diagram.txt file, but the file contains no line segments for that vertex.
In both cases, a NULL border is returned.
}
