% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{voronoiTreemap}
\alias{voronoiTreemap}
\title{Voronoi Treemap Algorithm}
\usage{
voronoiTreemap(coordX, coordY, Weights, minX, minY, maxX, maxY)
}
\arguments{
\item{coordX}{numeric vector of x coordinates}

\item{coordY}{numeric vector of y coordinates}

\item{Weights}{numeric vector of weights}
}
\description{
See \url{https://www.stat.auckland.ac.nz/~paul/Reports/VoronoiTreemap/voronoiTreeMap.html} for algorithm details.
}
