% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WeightedVoronoi}
\alias{WeightedVoronoi}
\title{Weighted Voronoi Diagram}
\usage{
WeightedVoronoi(coordX, coordY, Weights)
}
\arguments{
\item{coordX}{numeric vector of x coordinates}

\item{coordY}{numeric vector of y coordinates}

\item{Weights}{numeric vector of weights}
}
\description{
Produce a weighted Voronoi diagram. This is a wrapper for the CGAL library, see \url{http://doc.cgal.org/latest/Apollonius_graph_2/index.html} for algorithm details.
}
